+++ app/src/main/java/com/example/mvvm2/ui/main/MainActivity.kt
@@ -1,17 +1,11 @@
-package com.example.mvvm2.ui
+package com.example.mvvm2.ui.main
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
 import androidx.databinding.DataBindingUtil
 import androidx.lifecycle.ViewModelProvider
 import com.example.mvvm2.R
-import com.example.mvvm2.data.model.MovieResponse
-import com.example.mvvm2.data.repository.BaseRetrofit
-import com.example.mvvm2.data.repository.MovieRetrofit
 import com.example.mvvm2.databinding.ActivityMainBinding
-import retrofit2.Call
-import retrofit2.Response
 import javax.security.auth.callback.Callback
 import kotlin.math.log
