+++ app/src/main/res/layout/activity_main.xml
@@ -7,13 +7,13 @@
     <data>
         <variable
             name="vm"
-            type="com.example.mvvm2.ui.MainViewModel" />
+            type="com.example.mvvm2.ui.main.MainViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.MainActivity">
+        tools:context=".ui.main.MainActivity">
 
         <EditText
             android:id="@+id/main_search_edit"
