+++ app/src/main/res/layout/activity_movie.xml
@@ -3,14 +3,14 @@
     <data>
         <variable
             name="vm"
-            type="com.example.mvvm2.ui.MovieViewModel" />
+            type="com.example.mvvm2.ui.moive.MovieViewModel" />
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.MovieActivity">
+        tools:context=".ui.moive.MovieActivity">
 
         <ImageView
             android:id="@+id/movie_movie_image"
